TECH-1606 add support for task handler mode

+++ scripts/start_server.sh
@@ -10,17 +10,23 @@ source /home/ubuntu/.virtualenvs/ourrevolution/bin/virtualenvwrapper.sh
 # TODO: TECH-1294 debug error code
 workon ourrevolution || true
 
-# TODO: check if task machine mode is enabled
-echo "Restart celery"
-# Stop celery main process gracefully and auto-restart
-# TODO: TECH-1294 debug error code
-celery -A ourrevolution inspect stats | grep '"pid": ' | awk '{print $2}' | awk -F, '{print $1}' | xargs kill -TERM || true
+# Check if task handler is disabled
+if [ "$DISABLE_TASK_HANDLER" != "1" ]
+then
+  echo "Restart celery"
+  # Stop celery main process gracefully and auto-restart
+  # TODO: TECH-1294 debug error code
+  celery -A ourrevolution inspect stats | grep '"pid": ' | awk '{print $2}' | awk -F, '{print $1}' | xargs kill -TERM || true
+fi
 
-# TODO: check if web app mode is enabled
-echo "Start server"
-supervisorctl start gunicorn
+# Check if web app is disabled
+if [ "$DISABLE_WEB_APP" != "1" ]
+then
+  echo "Start web app"
+  supervisorctl start gunicorn
 
-echo "Purge Fastly"
-curl -XPOST -H "Fastly-Key:$FASTLY_API_KEY" https://api.fastly.com/service/$FASTLY_SERVICE_ID/purge_all
+  echo "Purge Fastly"
+  curl -XPOST -H "Fastly-Key:$FASTLY_API_KEY" https://api.fastly.com/service/$FASTLY_SERVICE_ID/purge_all
+fi
 
 echo "Server started"
